Read in fuel model constants from file
Previously, all constants from the fuel model were assigned as fuel model G, and all would have to be changed by hand if wanting to switch to a different fuel model. Now, with the additional input of a file containing all the fuel model constants, the only line of code to change is the assignment of the variable "fuel_mod".