new way of reading in z3 pduration coefficient
the function pjolly processes a string argument to either be a grid name or a file name, and from there reads in the proper z3 file. The current version has the "proper" lines commented out since it has not been officially contributed and therefore the filepaths that will eventually be right are not currently valid. Where it will be placed is not finalized, this is merely the first working version of it.