Check length of file inputted into fire index function
The function check_time_length has been added and implemented in fire index functions. This added the opt input variable to all the functions for the choice to suppress a warning about the length of time inputted. For KBDI the recommended length is 30 years (climatological average for rain). For all others it is 1 year. ERC is not included because some other changes were made before this was committed, will be committed separately.